requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrbox==0.1.5
azure-core==1.30.1
azure-storage-blob==12.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
deprecation==2.1.0
docopt==0.6.2
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.110.3
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.6.1
packaging==24.0
pdfmerge==1.0.0
pillow==10.3.0
postgrest==0.16.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pypdf==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
realtime==1.0.4
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
